Remove Plugin. Replace with Commands() and Expansions()
Also split out `types.go` into separate files by the area they cover
(services, auth, actions).
