+++ src/github.com/matrix-org/go-neb/services/echo/echo.go
@@ -1,10 +1,10 @@
 package services
 
 import (
-	"github.com/matrix-org/go-neb/matrix"
-	"github.com/matrix-org/go-neb/plugin"
-	"github.com/matrix-org/go-neb/types"
 	"strings"
+
+	"github.com/matrix-org/go-neb/matrix"
+	"github.com/matrix-org/go-neb/types"
 )
 
 type echoService struct {
@@ -16,14 +16,12 @@ type echoService struct {
 func (e *echoService) ServiceUserID() string { return e.serviceUserID }
 func (e *echoService) ServiceID() string     { return e.id }
 func (e *echoService) ServiceType() string   { return "echo" }
-func (e *echoService) Plugin(cli *matrix.Client, roomID string) plugin.Plugin {
-	return plugin.Plugin{
-		Commands: []plugin.Command{
-			plugin.Command{
-				Path: []string{"echo"},
-				Command: func(roomID, userID string, args []string) (interface{}, error) {
-					return &matrix.TextMessage{"m.notice", strings.Join(args, " ")}, nil
-				},
+func (e *echoService) Commands(cli *matrix.Client, roomID string) []types.Command {
+	return []types.Command{
+		types.Command{
+			Path: []string{"echo"},
+			Command: func(roomID, userID string, args []string) (interface{}, error) {
+				return &matrix.TextMessage{"m.notice", strings.Join(args, " ")}, nil
 			},
 		},
 	}
